refactor: move some helpers out of the service
