+++ src/apps/gift/app.ts
@@ -5,6 +5,7 @@ import config from "@config";
 
 import { hasNewModel, hasSchema } from "@core/middleware";
 import { wrapAsync } from "@core/utils";
+import { loginAndRedirect } from "@core/utils/member";
 
 import GiftService from "@core/services/GiftService";
 import MembersService from "@core/services/MembersService";
@@ -122,12 +123,7 @@ app.get(
           req.flash("warning", "gifts-already-activated");
           res.redirect("/login");
         } else {
-          MembersService.loginAndRedirect(
-            req,
-            res,
-            giftFlow.giftee,
-            "/profile/complete"
-          );
+          loginAndRedirect(req, res, giftFlow.giftee, "/profile/complete");
         }
       } else {
         next("route");
